GamePrefs::getInt() no longer exists; for uniformity, all numeric values will be stored as floats and then cast appropriately. Also, a few changes to ProgramLog to make it more efficient (or at least fewer SLOC).

+++ src/drawing/HUD/console/DrawConsole.cpp
@@ -178,13 +178,13 @@ void DrawConsole::draw() {
 				(unsigned int)((myElement->size.x  - containerPadding.x * 2.0f) / 2.0f * gameWindow->width),
 				(unsigned int)((myElement->size.y  - containerPadding.y * 2.0f) / 2.0f * gameWindow->height),
 				fontManager,
-				gamePrefs.getInt("fontStandardSize")
+				(int) gamePrefs.getFloat("fontStandardSize")
 		);
 		numLogLines = logStrings.size();
 	}
 
 	// draw the text
-	unsigned int fontSize = gamePrefs.getInt("fontStandardSize");
+	unsigned int fontSize = (unsigned int) gamePrefs.getFloat("fontStandardSize");
 
 	glUseProgram(program);
